% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{add.user}
\alias{add.user}
\title{Creates a new postgres db account}
\usage{
add.user(uname, pass = "nasa", con = db.options("db"))
}
\arguments{
\item{uname}{A \code{string} that indicates the username to be added.}

\item{con}{A \code{RPostgreSQL} connection.}
}
\description{
Builds a new postgres account with password `nasa` by default
}

