% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.options.R
\name{db.options}
\alias{db.options}
\title{Set or query db options}
\usage{
db.options(...)
}
\arguments{
\item{...}{A valid db parameter}
}
\description{
Set or query db options
}
\examples{
\dontrun{
# Establish connection
con = dbConnect(PostgreSQL(), user= "postgres", password="nasa", dbname="fia", host="localhost")

# Store connection in DB options
db.options("db" = con)
}
}

