% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{estimate.area.timberland}
\alias{estimate.area.timberland}
\title{Estimates the area of timberland}
\usage{
estimate.area.timberland(con = db.options("db"))
}
\arguments{
\item{con}{A \code{RPostgreSQL} connection.}
}
\value{
A \code{data.frame} containing the estimated area in timberland
}
\description{
Provides a numerical result of area of timberland in the U.S.
}

