% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{estimate.seedlings.live}
\alias{estimate.seedlings.live}
\title{Estimates the area of Seedlings on Forest Land}
\usage{
estimate.seedlings.live(con = db.options("db"))
}
\arguments{
\item{con}{A \code{RPostgreSQL} connection.}
}
\value{
A \code{data.frame} containing the estimated area.
}
\description{
Provides a count of numerical values on forest level.
}

