% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{estimate.trees.forest}
\alias{estimate.trees.forest}
\title{Estimates the area of Live Trees on Forest Land}
\usage{
estimate.trees.forest(con = db.options("db"))
}
\arguments{
\item{con}{A \code{RPostgreSQL} connection.}
}
\value{
A \code{data.frame} containing the estimated area.
}
\description{
Provides a count of numerical values on forest level.
}

