% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiadb_package.R
\docType{package}
\name{fiadb-package}
\alias{fiadb-package}
\title{fiadb Package}
\description{
fiadb
}
\details{
\tabular{ll}{
Package: \tab fiadb\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-05-26\cr
License: \tab MIT + file LICENSE\cr
}
Some details should be added
}
\examples{
# Fill in at a later time
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Bo Li \email{libo@illinois.edu}

Bo Li \email{libo@illinois.edu}
}

