% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get.state}
\alias{get.state}
\title{Check if entry is a valid state}
\usage{
get.state(state)
}
\arguments{
\item{state}{a \code{string} that is a valid two letter state code (e.g. Illinois is "IL").}
}
\value{
returns state code: otherwise, it'll kill the process.
}
\description{
Searches list for acronym / state code.
}

