% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{state.dbh.extended}
\alias{state.dbh.extended}
\title{DBH Information for Trees in Given State}
\usage{
state.dbh.extended(state, con = db.options("db"))
}
\arguments{
\item{state}{A \code{int} or \code{string} that contains either the state ID or}

\item{con}{A \code{RPostgreSQL} connection.}
}
\value{
A \code{data.frame} containing tree dbh information in states.
}
\description{
Trees in a given state win!
}

