% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{state.species}
\alias{state.species}
\title{Obtain a list of species in a given state}
\usage{
state.species(state, con = db.options("db"))
}
\arguments{
\item{state}{A \code{int} or \code{string} that contains either the state ID or}

\item{con}{A \code{RPostgreSQL} connection.}
}
\value{
A \code{data.frame} containing species in states.
}
\description{
Provides a list of species that exist in a given state
}
\details{
Follows the state codes listed in here:
}

